Sphinx documentation: Added documentation main package ('ospx')
