+++ docs/source/api.rst
@@ -5,4 +5,3 @@ API Documentation
    :maxdepth: 4
 
    ospx
-   ospx.fmi
