+++ docs/source/ospx.graph.rst
@@ -1,0 +1,7 @@
+ospx.graph module
+=================
+
+.. automodule:: ospx.graph
+   :members:
+   :undoc-members:
+   :show-inheritance:
