+++ docs/source/ospx.ospSimulationCase.rst
@@ -1,0 +1,7 @@
+ospx.ospSimulationCase module
+=============================
+
+.. automodule:: ospx.ospSimulationCase
+   :members:
+   :undoc-members:
+   :show-inheritance:
