+++ src/ospx/__init__.py
@@ -1,6 +1,8 @@
-from .simulation import Simulation                              # noqa: F401
-from .connector import Connector                                # noqa: F401
-from .connection import Endpoint, Connection                    # noqa: F401
-from .component import Component                                # noqa: F401
-from .systemStructure import SystemStructure                    # noqa: F401
-from .ospCaseBuilder import OspCaseBuilder, OspSimulationCase   # noqa: F401
+from .simulation import Simulation                  # noqa: F401
+from .connector import Connector                    # noqa: F401
+from .connection import Endpoint, Connection        # noqa: F401
+from .component import Component                    # noqa: F401
+from .systemStructure import SystemStructure        # noqa: F401
+from .ospSimulationCase import OspSimulationCase    # noqa: F401
+from .graph import Graph                            # noqa: F401
+from .ospCaseBuilder import OspCaseBuilder          # noqa: F401
