+++ tests/test_dicts/test_caseDict_imported_test_graph
@@ -1,7 +1,6 @@
 /*---------------------------------*- C++ -*----------------------------------*\
 filetype dictionary; coding utf-8; version 0.1; local --; purpose --;
 \*----------------------------------------------------------------------------*/
-#include '.\paramDict'
 
 _environment
 {
